Move project data into separate file and remove commented code
