+++ index.html
@@ -11,7 +11,7 @@
   <link href='https://fonts.googleapis.com/css?family=Crete Round' rel='stylesheet'>
   <link href='https://fonts.googleapis.com/css?family=Poppins' rel='stylesheet'>
   <link href='https://fonts.googleapis.com/css?family=Inter' rel='stylesheet'>
-  <script defer src="index.js"></script>
+  <script type="module" defer src="index.js"></script>
 </head>
 <body>
   <main data-main>
@@ -82,136 +82,6 @@
       <img src="images/Desktop-healine-image.png" alt="" class="desktop-healine-image flex">
     </section>
     <section id="works-section">
-      
-      
-      <!-- <div class="works-title-block">
-        <h3 class="title flex">My Recent Works</h3>
-        <div class="divider-desktop">
-          <img class="desktop-divider" src="images/desktop-divider.png" alt="">
-        </div>
-      </div>
-      <div class="crossline"><img src="images/Vector (2).png" alt=""></div>
-      <div class="works-card">
-        <div class="yoga-placeholder"><img src="images/Img Placeholder.png" alt=""></div>
-        <h3 id="bold-text" class="flex">Multi-Post Stories</h3>
-        <p id="works-descriptiofn">A daily selection of privately personalized reads; no accounts or sign-ups required.
-          has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a standard
-          dummy text.</p>
-        <ul id="competencies" class="flex">
-          <li>CSS</li>
-          <li>HTML</li>
-          <li>JAVASCRIPT</li>
-        </ul>
-        <button type="button" class="btn-type-a flex">See project</button>
-      </div>
-      <div class="works-card-desktop flex">
-        <img src="images/Img Placeholder.png" alt="" class="project-cards flex">
-        <div class="side-text">
-          <h3 id="bold-text">Multi-Post Stories</h3>
-          <p id="works-description">A daily selection of privately personalized reads; no accounts or sign-ups required. has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a standard dummy text.</p>
-          <ul id="competencies" class="flex">
-            <li>CSS</li>
-            <li>HTML</li>
-            <li>Bootstrap</li>
-            <li>Ruby</li>
-          </ul>
-          <button type="button" class="btn-type-a">See project</button>
-        </div>
-      </div>
-      <div class="grid-container">
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(2).png" alt="" class="project-cards flex">
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(3).png" alt="" class="project-cards flex">
-          <div class="image-text flex">
-            <div>
-              <h3 class="bold-text flex">Profesional Art Printing Data</h3>
-            </div>
-            <p class="works-description">A daily selection of privately personalized reads; no accounts or sign-ups
-              required. Has been the industry's standard.</p>
-            <ul class="competencies flex">
-              <li>CSS</li>
-              <li>HTML</li>
-              <li>JAVASCRIPT</li>
-            </ul>
-          </div>
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(4).png" alt="" class="project-cards flex">
-          <div class="image-text">
-            <div>
-              <h3 class="bold-text">Profesional Art Printing Data</h3>
-            </div>
-            <p class="works-description">A daily selection of privately personalized reads; no accounts or sign-ups
-              required. Has been the industry's standard.</p>
-            <ul class="competencies flex">
-              <li>CSS</li>
-              <li>HTML</li>
-              <li>JAVASCRIPT</li>
-            </ul>
-          </div>
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards flex">
-          <div class="image-text">
-            <div>
-              <h3 class="bold-text">Profesional Art Printing Data</h3>
-            </div>
-            <p class="works-description">A daily selection of privately personalized reads; no accounts or sign-ups
-              required. Has been the industry's standard.</p>
-            <ul class="competencies flex">
-              <li>CSS</li>
-              <li>HTML</li>
-              <li>JAVASCRIPT</li>
-            </ul>
-          </div>
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(3).png" alt="" class="project-cards flex">
-          <div class="image-text">
-            <div>
-              <h3 class="bold-text">Profesional Art Printing Data</h3>
-            </div>
-            <p class="works-description">A daily selection of privately personalized reads; no accounts or sign-ups
-              required. Has been the industry's standard.</p>
-            <ul class="competencies flex">
-              <li>CSS</li>
-              <li>HTML</li>
-              <li>JAVASCRIPT</li>
-            </ul>
-          </div>
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-        <div class="grid-item">
-          <img src="images/Img-Placeholder(1).png" alt="" class="project-cards-mobile">
-          <img src="images/Img-Placeholder(4).png" alt="" class="project-cards flex">
-          <div class="image-text">
-            <div>
-              <h3 class="bold-text">Profesional Art Printing Data</h3>
-            </div>
-            <p class="works-description">A daily selection of privately personalized reads; no accounts or sign-ups
-              required. Has been the industry's standard.</p>
-            <ul class="competencies flex">
-              <li>CSS</li>
-              <li>HTML</li>
-              <li>JAVASCRIPT</li>
-            </ul>
-          </div>
-          <div><button type="button" class="btn-type-b flex">See project</button></div>
-        </div>
-      </div> -->
-
-
     </section>
     <section class="about-me-section flex" id="about-section">
       <h2 class="about-me-title flex">About me</h2>
@@ -327,7 +197,7 @@
         </li>
       </ul>
     </footer>
-    <div class="footer-divider flex"><div></div></div> 
+    <div class="footer-divider flex"><div></div></div>
   </main>
 </body>
 </html>
