+++ index.js
@@ -1,3 +1,7 @@
+import projectData from "./projects_data.js";
+
+console.log(projectData);
+
 const mobileVersionContainer = document.querySelector('[data-main]');
 const menuButton = document.querySelector('[data-menu]');
 
@@ -119,136 +123,6 @@ worksDesktop();
 
 const cardsContainer = document.createElement('div');
 cardsContainer.className = 'grid-container';
-
-const projectData = [
-  {
-    name: 'To-do list',
-    description: 'This To Do list is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete when you are done. It also allows you to edit and delete a tasks as you want. It was built using Webpack, HTML, CSS',
-    'Featured image': {
-      mobile: {
-        src: 'images/Screenshot_todo-list.png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Screenshot_todo-list.png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': 'https://forloop94.github.io/To-do-list/dist/',
-    'Link to source': 'https://github.com/forLoop94/To-do-list',
-  },
-
-  {
-    name: 'Conference Page',
-    description: 'African Fiction Festival 2023 is a website that displays information about the African Fiction Festival that will be holding for the first time in Abuja, Nigeria. It was built using HTML, CSS and Javascript.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Screenshot_AFF-mobile.png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Screenshot_AFF-desktop.png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': 'https://forloop94.github.io/Conference_Page/',
-    'Link to source': 'https://github.com/forLoop94/Conference_Page',
-  },
-  {
-    name: 'Project Three',
-    description: 'You need to update the main page so that the projects section is created dynamically using the information stored in that JavaScript object. Remember, all of the HTML in that section is created when the page loads.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Img-Placeholder(4).png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': '#',
-    'Link to source': '#',
-  },
-  {
-    name: 'Project Four',
-    description: 'You need to update the main page so that the projects section is created dynamically using the information stored in that JavaScript object. Remember, all of the HTML in that section is created when the page loads.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': '#',
-    'Link to source': '#',
-  },
-  {
-    name: 'Project Five',
-    description: 'You need to update the main page so that the projects section is created dynamically using the information stored in that JavaScript object. Remember, all of the HTML in that section is created when the page loads.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Img-Placeholder(3).png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': '#',
-    'Link to source': '#',
-  },
-  {
-    name: 'Project Six',
-    description: 'You need to update the main page so that the projects section is created dynamically using the information stored in that JavaScript object. Remember, all of the HTML in that section is created when the page loads.',
-    'Featured image': {
-      mobile: {
-        src: 'images/Img-Placeholder(1).png',
-        alt: '',
-      },
-      desktop: {
-        src: 'images/Img-Placeholder(4).png',
-        alt: '',
-      },
-    },
-    technologies: {
-      tech1: 'CSS',
-      tech2: 'HTML',
-      tech3: 'Javascript',
-    },
-    'Link to live version': '#',
-    'Link to source': '#',
-  },
-];
 
 function recurringCards(arr) {
   for (let i = 0; i < arr.length; i += 1) {
